requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.24.0
plotly>=5.15.0
chardet>=5.0.0
paramiko>=3.0.0
gspread>=5.0.0
oauth2client>=4.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
apscheduler>=3.10.0
pytz>=2023.3
croniter>=2.0.0
tenacity>=8.2.0
backoff>=2.2.0